refactor: split library episode handlers into submodule
Extract episode-file management handlers (delete, cancel, grab history,
mark-failed, progress, JSON snapshot) plus EpisodeProgress into
handlers/library/episodes.rs. The shared helpers resolve_series_context,
resolve_tracked_series, build_episodes, and run_auto_search_targets gain
pub(super) so the submodule can reuse them. library/mod.rs drops from
5541 to 4739 lines; episodes.rs is 535.

main.rs OpenApi references bump to handlers::library::episodes::* for
the same utoipa/__path_X reason as the prior settings split.

No behavior change. cargo test still 480 passes.
