+++ src/main.rs
@@ -58,12 +58,12 @@ use services::{
         handlers::library::interactive_search_batches,
         handlers::library::grab_interactive_result,
         handlers::library::grab_batch_result,
-        handlers::library::delete_episode_file,
-        handlers::library::cancel_pending_episode,
-        handlers::library::get_episode_grab_history,
-        handlers::library::mark_episode_failed,
-        handlers::library::episode_download_progress,
-        handlers::library::series_episodes_json,
+        handlers::library::episodes::delete_episode_file,
+        handlers::library::episodes::cancel_pending_episode,
+        handlers::library::episodes::get_episode_grab_history,
+        handlers::library::episodes::mark_episode_failed,
+        handlers::library::episodes::episode_download_progress,
+        handlers::library::episodes::series_episodes_json,
         // Search
         handlers::search::search_page_api,
         handlers::search::grab_release,
